+++ example/pubspec.dart
@@ -2,13 +2,11 @@
 // source code
 // is governed by a BSD-style license that can be found in the LICENSE file.
 
-import 'dart:io';
-
 import 'package:pubspec2/pubspec2.dart';
 
 void main() async {
   // specify the directory
-  final myDirectory = Directory('myDir');
+  const myDirectory = 'myDir';
 
   // load pubSpec
   final pubSpec = await PubSpec.load(myDirectory);