+++ test/pubspec_test.dart
@@ -1,8 +1,7 @@
 @Skip('not a real test')
 library;
 
-import 'dart:io';
-
+import 'package:dcli_core/dcli_core.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:pubspec2/pubspec2.dart';
 import 'package:test/test.dart';
@@ -13,5 +12,7 @@ void main() async {
     'fred': HostedReference(VersionRange(min: Version(1, 2, 3)))
   });
 
-  await pubSpec.save(await Directory.systemTemp.createTemp('delme'));
+  await withTempDir((tempDir) async {
+    await pubSpec.save(tempDir);
+  });
 }